+++ deploy/kubernetes/config-map.yaml
@@ -10,7 +10,7 @@ data:
   config.json: |-
     {
       "simplybk": {
-        "uuid": "79276661-5f8a-405d-ab6d-651b88326206",
-        "ip": "18.218.243.112"
+        "uuid": "cdcd53dd-f8d2-4978-9063-a4f0189628b1",
+        "ip": "35.170.65.79"
       }
     }
