+++ deploy/kubernetes/secret.yaml
@@ -16,6 +16,6 @@ stringData:
   secret.json: |-
     {
       "simplybk": {
-        "secret": "8K9BVlNfYo6aiLamu1c2"
+        "secret": "azZJBC0fYjJMRP0hC0NT"
       }
     }
