+++ e2e/e2e.go
@@ -14,12 +14,8 @@ import (
 	"k8s.io/kubernetes/test/e2e/storage/podlogs"
 )
 
-var runXPU bool
-
 func init() {
 	klog.SetOutput(ginkgo.GinkgoWriter)
-
-	flag.BoolVar(&runXPU, "xpu", true, "Run XPU tests")
 
 	if os.Getenv("KUBECONFIG") == "" {
 		kubeConfigPath := filepath.Join(os.Getenv("HOME"), ".kube", "config")
