--- e2e/sma-nvmf.yaml
@@ -1,19 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright (c) Arm Limited and Contributors
-# Copyright (c) Intel Corporation
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: spdkcsi-nodeservercm
-data:
-  nodeserver-config.json: |-
-    {
-      "xpuList": [
-        {
-          "name": "xPU0",
-          "targetType": "xpu-sma-nvmftcp",
-          "targetAddr": "127.0.0.1:5114"
-        }
-      ]
-    }
