--- e2e/sma-virtioblk.yaml
@@ -1,20 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright (c) Arm Limited and Contributors
-# Copyright (c) Intel Corporation
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: spdkcsi-nodeservercm
-data:
-  nodeserver-config.json: |-
-    {
-      "xpuList": [
-        {
-          "name": "xPU0",
-          "targetType": "xpu-sma-virtioblk",
-          "targetAddr": "127.0.0.1:5114"
-        }
-      ],
-      "kvmPciBridges": 2
-    }
